chore: split webpack configuration
removed write-file-webpack-plugin
