+++ config-overrides/template.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+    <script src="/polyfill/browser-polyfill.min.js"></script>
+    <script src="/firefoxFix.js"></script>
+</head>
+<body></body>
+</html>
